[eslint-plugin] separate cssProperties from valid-styles
