+++ packages/@stylexjs/eslint-plugin/src/stylex-valid-styles.js
@@ -9,7 +9,6 @@
 
 'use strict';
 
-import namedColors from './reference/namedColors';
 import getDistance from './utils/getDistance';
 import isWhiteSpaceOrEmpty from './utils/isWhiteSpaceOrEmpty';
 import type {
@@ -32,22 +31,24 @@ import type {
 } from 'estree';
 import micromatch from 'micromatch';
 /*:: import { Rule } from 'eslint'; */
-import isCSSVariable from './rules/isCSSVariable';
 import makeLiteralRule from './rules/makeLiteralRule';
-import makeRangeRule from './rules/makeRangeRule';
-import makeRegExRule from './rules/makeRegExRule';
 import isString from './rules/isString';
-import isHexColor from './rules/isHexColor';
 import makeUnionRule from './rules/makeUnionRule';
 import isNumber from './rules/isNumber';
-import isPercentage from './rules/isPercentage';
 import isAnimationName from './rules/isAnimationName';
-import isAbsoluteLength from './rules/isAbsoluteLength';
-import isRelativeLength from './rules/isRelativeLength';
 import isStylexDefineVarsToken from './rules/isStylexDefineVarsToken';
-import { borderSplitter } from './utils/split-css-value';
 import evaluate from './utils/evaluate';
 import resolveKey from './utils/resolveKey';
+import {
+  CSSPropertyKeys,
+  CSSProperties,
+  CSSPropertyReplacements,
+  pseudoElements,
+  convertToStandardProperties,
+  pseudoClassesAndAtRules,
+  allModifiers,
+  all,
+} from './reference/cssProperties';
 
 export type Variables = $ReadOnlyMap<string, Expression | 'ARG'>;
 export type RuleCheck = (
@@ -67,2035 +68,6 @@ export type RuleResponse = void | {
 const showError =
   (message: string): RuleCheck =>
   () => ({ message });
-
-const isStringOrNumber = makeUnionRule(isString, isNumber);
-
-const isNamedColor = makeUnionRule(
-  ...Array.from(namedColors).map((color) => makeLiteralRule(color)),
-);
-
-const isLength = makeUnionRule(isAbsoluteLength, isRelativeLength);
-
-const isNonNumericString: RuleCheck = (node: Node): RuleResponse => {
-  if (node.type === 'Literal' && typeof node.value === 'string') {
-    if (/^[-+]?(?:\d+|\d*\.\d+)$/.test(node.value)) {
-      return {
-        message: 'a non-numeric string',
-      };
-    }
-  }
-  return undefined;
-};
-
-// NOTE: converted from Flow types to function calls using this
-// https://astexplorer.net/#/gist/87e64b378349f13e885f9b6968c1e556/4b4ff0358de33cf86b8b21d29c17504d789babf9
-const all: RuleCheck = makeUnionRule(
-  makeLiteralRule(null),
-  makeLiteralRule('initial'),
-  makeLiteralRule('inherit'),
-  makeLiteralRule('unset'),
-  makeLiteralRule('revert'),
-);
-
-const length = makeUnionRule(isNumber, isNonNumericString);
-
-const color = makeUnionRule(isString, isNamedColor, isHexColor);
-const width = makeUnionRule(
-  isNonNumericString,
-  isNumber,
-  makeLiteralRule('available'),
-  makeLiteralRule('min-content'),
-  makeLiteralRule('max-content'),
-  makeLiteralRule('fit-content'),
-  makeLiteralRule('auto'),
-  isLength,
-  isPercentage,
-);
-const borderWidth = makeUnionRule(
-  isNumber,
-  makeLiteralRule('thin'),
-  makeLiteralRule('medium'),
-  makeLiteralRule('thick'),
-  isNonNumericString,
-  isLength,
-);
-const lengthPercentage = isStringOrNumber;
-const borderImageSource = makeUnionRule(makeLiteralRule('none'), isString);
-const time = isString;
-const animationDirection = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('reverse'),
-  makeLiteralRule('alternate'),
-  makeLiteralRule('alternate-reverse'),
-);
-const animationFillMode = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('forwards'),
-  makeLiteralRule('backwards'),
-  makeLiteralRule('both'),
-);
-const animationIterationCount = makeUnionRule(
-  makeLiteralRule('infinite'),
-  isNumber,
-);
-
-const animationPlayState = makeUnionRule(
-  makeLiteralRule('running'),
-  makeLiteralRule('paused'),
-);
-const animationTimingFunction = makeUnionRule(
-  makeLiteralRule('ease'),
-  makeLiteralRule('linear'),
-  makeLiteralRule('ease-in'),
-  makeLiteralRule('ease-out'),
-  makeLiteralRule('ease-in-out'),
-  makeLiteralRule('step-start'),
-  makeLiteralRule('step-end'),
-  isString,
-);
-const attachment = makeUnionRule(
-  makeLiteralRule('scroll'),
-  makeLiteralRule('fixed'),
-  makeLiteralRule('local'),
-);
-const blendMode = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('multiply'),
-  makeLiteralRule('screen'),
-  makeLiteralRule('overlay'),
-  makeLiteralRule('darken'),
-  makeLiteralRule('lighten'),
-  makeLiteralRule('color-dodge'),
-  makeLiteralRule('color-burn'),
-  makeLiteralRule('hard-light'),
-  makeLiteralRule('soft-light'),
-  makeLiteralRule('difference'),
-  makeLiteralRule('exclusion'),
-  makeLiteralRule('hue'),
-  makeLiteralRule('saturation'),
-  makeLiteralRule('color'),
-  makeLiteralRule('luminosity'),
-);
-const bgSize = makeUnionRule(
-  isString,
-  makeLiteralRule('cover'),
-  makeLiteralRule('contain'),
-);
-const boxAlign = makeUnionRule(
-  makeLiteralRule('start'),
-  makeLiteralRule('center'),
-  makeLiteralRule('end'),
-  makeLiteralRule('baseline'),
-  makeLiteralRule('stretch'),
-);
-const repeatStyle = makeUnionRule(
-  makeLiteralRule('repeat-x'),
-  makeLiteralRule('repeat-y'),
-  isString,
-);
-const backgroundPosition = makeUnionRule(
-  isString,
-  makeLiteralRule('top'),
-  makeLiteralRule('bottom'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('center'),
-);
-const backgroundPositionX = makeUnionRule(
-  isString,
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('center'),
-);
-const backgroundPositionY = makeUnionRule(
-  isString,
-  makeLiteralRule('top'),
-  makeLiteralRule('bottom'),
-  makeLiteralRule('center'),
-);
-const borderImageOutset = isString;
-const borderImageRepeat = makeUnionRule(
-  isString,
-  makeLiteralRule('stretch'),
-  makeLiteralRule('repeat'),
-  makeLiteralRule('round'),
-  makeLiteralRule('space'),
-);
-const borderImageWidth = isString;
-const borderImageSlice = makeUnionRule(
-  isStringOrNumber,
-  makeLiteralRule('fill'),
-);
-const box = makeUnionRule(
-  makeLiteralRule('border-box'),
-  makeLiteralRule('padding-box'),
-  makeLiteralRule('content-box'),
-);
-const brStyle = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('hidden'),
-  makeLiteralRule('dotted'),
-  makeLiteralRule('dashed'),
-  makeLiteralRule('solid'),
-  makeLiteralRule('double'),
-  makeLiteralRule('groove'),
-  makeLiteralRule('ridge'),
-  makeLiteralRule('inset'),
-  makeLiteralRule('outset'),
-);
-const CSSCursor = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('default'),
-  makeLiteralRule('none'),
-  makeLiteralRule('context-menu'),
-  makeLiteralRule('help'),
-  makeLiteralRule('pointer'),
-  makeLiteralRule('progress'),
-  makeLiteralRule('wait'),
-  makeLiteralRule('cell'),
-  makeLiteralRule('crosshair'),
-  makeLiteralRule('text'),
-  makeLiteralRule('vertical-text'),
-  makeLiteralRule('alias'),
-  makeLiteralRule('copy'),
-  makeLiteralRule('move'),
-  makeLiteralRule('no-drop'),
-  makeLiteralRule('not-allowed'),
-  makeLiteralRule('e-resize'),
-  makeLiteralRule('n-resize'),
-  makeLiteralRule('ne-resize'),
-  makeLiteralRule('nw-resize'),
-  makeLiteralRule('s-resize'),
-  makeLiteralRule('se-resize'),
-  makeLiteralRule('sw-resize'),
-  makeLiteralRule('w-resize'),
-  makeLiteralRule('ew-resize'),
-  makeLiteralRule('ns-resize'),
-  makeLiteralRule('nesw-resize'),
-  makeLiteralRule('nwse-resize'),
-  makeLiteralRule('col-resize'),
-  makeLiteralRule('row-resize'),
-  makeLiteralRule('all-scroll'),
-  makeLiteralRule('zoom-in'),
-  makeLiteralRule('zoom-out'),
-  makeLiteralRule('grab'),
-  makeLiteralRule('grabbing'),
-  makeLiteralRule('-webkit-grab'),
-  makeLiteralRule('-webkit-grabbing'),
-);
-const relativeSize = makeUnionRule(
-  makeLiteralRule('larger'),
-  makeLiteralRule('smaller'),
-);
-const emptyCells = makeUnionRule(
-  makeLiteralRule('show'),
-  makeLiteralRule('hide'),
-);
-const filter = makeUnionRule(makeLiteralRule('none'), isString);
-// const flex = makeUnionRule(makeLiteralRule('none'), isString, isNumber);
-const flexBasis = makeUnionRule(makeLiteralRule('content'), isNumber, isString);
-const flexDirection = makeUnionRule(
-  makeLiteralRule('row'),
-  makeLiteralRule('row-reverse'),
-  makeLiteralRule('column'),
-  makeLiteralRule('column-reverse'),
-);
-const flexWrap = makeUnionRule(
-  makeLiteralRule('nowrap'),
-  makeLiteralRule('wrap'),
-  makeLiteralRule('wrap-reverse'),
-);
-const flexGrow = isStringOrNumber;
-const flexShrink = isStringOrNumber;
-const flexFlow = makeUnionRule(flexDirection, flexWrap);
-const float = makeUnionRule(
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('none'),
-  makeLiteralRule('inline-start'),
-  makeLiteralRule('inline-end'),
-);
-const absoluteSize = makeUnionRule(
-  makeLiteralRule('xx-small'),
-  makeLiteralRule('x-small'),
-  makeLiteralRule('small'),
-  makeLiteralRule('medium'),
-  makeLiteralRule('large'),
-  makeLiteralRule('x-large'),
-  makeLiteralRule('xx-large'),
-);
-const fontFamily = isString;
-const gridLine = makeUnionRule(makeLiteralRule('auto'), isString);
-const gridTemplate = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('subgrid'),
-  isString,
-);
-const gridTemplateAreas = makeUnionRule(makeLiteralRule('none'), isString);
-const trackBreadth = makeUnionRule(
-  lengthPercentage,
-  isString,
-  makeLiteralRule('min-content'),
-  makeLiteralRule('max-content'),
-  makeLiteralRule('auto'),
-);
-const listStyleType = makeUnionRule(isString, makeLiteralRule('none'));
-const trackSize = makeUnionRule(trackBreadth, isString);
-const borderStyle = brStyle;
-const columnRuleColor = color;
-const columnRuleStyle = brStyle;
-const columnRuleWidth = borderWidth;
-const columnRule = makeUnionRule(
-  columnRuleWidth,
-  columnRuleStyle,
-  columnRuleColor,
-);
-const shapeBox = makeUnionRule(box, makeLiteralRule('margin-box'));
-const geometryBox = makeUnionRule(
-  shapeBox,
-  makeLiteralRule('fill-box'),
-  makeLiteralRule('stroke-box'),
-  makeLiteralRule('view-box'),
-);
-const maskReference = makeUnionRule(makeLiteralRule('none'), isString);
-const compositeOperator = makeUnionRule(
-  makeLiteralRule('add'),
-  makeLiteralRule('subtract'),
-  makeLiteralRule('intersect'),
-  makeLiteralRule('exclude'),
-);
-const maskingMode = makeUnionRule(
-  makeLiteralRule('alpha'),
-  makeLiteralRule('luminance'),
-  makeLiteralRule('match-source'),
-);
-const maskLayer = makeUnionRule(
-  maskReference,
-  maskingMode,
-  isString,
-  repeatStyle,
-  geometryBox,
-  compositeOperator,
-);
-
-const alignBase = makeUnionRule(
-  'normal',
-  'stretch',
-  'center',
-  'start',
-  'end',
-  'flex-start',
-  'flex-end',
-  'baseline',
-  'first baseline',
-  'last baseline',
-  'safe center',
-  'unsafe center',
-);
-
-const alignContent = makeUnionRule(
-  alignBase,
-  'space-between',
-  'space-around',
-  'space-evenly',
-);
-
-const alignItems = makeUnionRule(alignBase, 'self-start', 'self-end');
-
-const alignSelf = makeUnionRule(alignBase, 'auto', 'self-start', 'self-end');
-const appearance = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('none'),
-  makeLiteralRule('textfield'),
-);
-const backdropFilter = makeUnionRule(makeLiteralRule('none'), isString);
-const backfaceVisibility = makeUnionRule(
-  makeLiteralRule('visible'),
-  makeLiteralRule('hidden'),
-);
-// type background = string | finalBgLayer;
-const backgroundAttachment = attachment;
-const backgroundBlendMode = blendMode;
-const backgroundClip = makeUnionRule(
-  'border-box',
-  'padding-box',
-  'content-box',
-  'text',
-);
-const backgroundColor = color;
-const backgroundImage = makeUnionRule(makeLiteralRule('none'), isString);
-const backgroundOrigin = box;
-const backgroundRepeat = repeatStyle;
-const backgroundSize = bgSize;
-const blockSize = width;
-const quotedString = (val: number | string) =>
-  typeof val === 'string' ? `'${val}'` : val;
-const border =
-  (suffix: string = ''): RuleCheck =>
-  (node: Expression | Pattern, _variables?: Variables, prop?: Property) => {
-    const response: $NonMaybeType<RuleResponse> = {
-      message: `The 'border${suffix}' property is not supported. Use the 'border${suffix}Width', 'border${suffix}Style' and 'border${suffix}Color' properties instead.`,
-    };
-    if (node.type !== 'Literal' || prop == null) {
-      return response;
-    }
-    if (typeof node.value === 'number') {
-      response.suggest = {
-        desc: `Replace 'border${suffix}' set to a number with 'border${suffix}Width' instead?`,
-        fix: (fixer: Rule.RuleFixer): Rule.Fix | null => {
-          return fixer.replaceText(
-            prop,
-            `border${suffix}Width: ${String(node.value)}`,
-          );
-        },
-      };
-    }
-    if (typeof node.value === 'string') {
-      const [width, style, color] = borderSplitter(node.value);
-      if (width != null || style != null || color != null) {
-        response.suggest = {
-          desc: `Replace 'border${suffix}' with 'border${suffix}Width', 'border${suffix}Style' and 'border${suffix}Color' instead?`,
-          fix: (fixer: Rule.RuleFixer): Rule.Fix | null => {
-            const newRules = [];
-            if (width != null) {
-              newRules.push(`border${suffix}Width: ${quotedString(width)}`);
-            }
-            if (style != null) {
-              newRules.push(`border${suffix}Style: ${quotedString(style)}`);
-            }
-            if (color != null) {
-              newRules.push(`border${suffix}Color: ${quotedString(color)}`);
-            }
-            return fixer.replaceText(prop, newRules.join(',\n    '));
-          },
-        };
-      }
-    }
-    return response;
-  };
-const borderBottomStyle = brStyle;
-const borderBottomWidth = borderWidth;
-const borderCollapse = makeUnionRule(
-  makeLiteralRule('collapse'),
-  makeLiteralRule('separate'),
-);
-const borderColor = color;
-const borderImage = makeUnionRule(
-  borderImageSource,
-  borderImageSlice,
-  isString,
-  borderImageRepeat,
-);
-const borderLeftColor = color;
-const borderLeftStyle = brStyle;
-const borderLeftWidth = borderWidth;
-const borderSpacing = isStringOrNumber;
-const borderTopStyle = brStyle;
-const borderTopWidth = borderWidth;
-const boxDecorationBreak = makeUnionRule(
-  makeLiteralRule('slice'),
-  makeLiteralRule('clone'),
-);
-const boxDirection = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('reverse'),
-);
-const boxFlex = isStringOrNumber;
-const boxFlexGroup = isStringOrNumber;
-const boxLines = makeUnionRule(
-  makeLiteralRule('single'),
-  makeLiteralRule('multiple'),
-);
-const boxOrdinalGroup = isStringOrNumber;
-const boxOrient = makeUnionRule(
-  makeLiteralRule('horizontal'),
-  makeLiteralRule('vertical'),
-  makeLiteralRule('inline-axis'),
-  makeLiteralRule('block-axis'),
-);
-const boxShadow = makeUnionRule(makeLiteralRule('none'), isString);
-const boxSizing = makeUnionRule(
-  makeLiteralRule('content-box'),
-  makeLiteralRule('border-box'),
-);
-const boxSuppress = makeUnionRule(
-  makeLiteralRule('show'),
-  makeLiteralRule('discard'),
-  makeLiteralRule('hide'),
-);
-const breakBeforeOrAfter = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('avoid'),
-  makeLiteralRule('avoid-page'),
-  makeLiteralRule('page'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('recto'),
-  makeLiteralRule('verso'),
-  makeLiteralRule('avoid-column'),
-  makeLiteralRule('column'),
-  makeLiteralRule('avoid-region'),
-  makeLiteralRule('region'),
-);
-const breakInside = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('avoid'),
-  makeLiteralRule('avoid-page'),
-  makeLiteralRule('avoid-column'),
-  makeLiteralRule('avoid-region'),
-);
-const captionSide = makeUnionRule(
-  makeLiteralRule('top'),
-  makeLiteralRule('bottom'),
-  makeLiteralRule('block-start'),
-  makeLiteralRule('block-end'),
-  makeLiteralRule('inline-start'),
-  makeLiteralRule('inline-end'),
-);
-const clear = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('both'),
-  makeLiteralRule('inline-start'),
-  makeLiteralRule('inline-end'),
-);
-const clip = makeUnionRule(isString, makeLiteralRule('auto'));
-const clipPath = makeUnionRule(isString, makeLiteralRule('none'));
-const columnCount = makeUnionRule(isNumber, isString, makeLiteralRule('auto'));
-const columnFill = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('balance'),
-);
-const columnGap = makeUnionRule(
-  isNumber,
-  isNonNumericString,
-  makeLiteralRule('normal'),
-);
-const columnSpan = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('all'),
-);
-const columnWidth = makeUnionRule(
-  isNumber,
-  isNonNumericString,
-  makeLiteralRule('auto'),
-);
-const columns = makeUnionRule(columnWidth, columnCount);
-const contain = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('strict'),
-  makeLiteralRule('content'),
-  isString,
-);
-const content = isString;
-const counterIncrement = makeUnionRule(isString, makeLiteralRule('none'));
-const counterReset = makeUnionRule(isString, makeLiteralRule('none'));
-const cursor = CSSCursor;
-const direction = makeUnionRule(makeLiteralRule('ltr'), makeLiteralRule('rtl'));
-const display = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('inline'),
-  makeLiteralRule('block'),
-  makeLiteralRule('flow-root'),
-  makeLiteralRule('list-item'),
-  makeLiteralRule('inline-list-item'),
-  makeLiteralRule('inline-block'),
-  makeLiteralRule('inline-table'),
-  makeLiteralRule('table'),
-  makeLiteralRule('table-cell'),
-  makeLiteralRule('table-column'),
-  makeLiteralRule('table-column-group'),
-  makeLiteralRule('table-footer-group'),
-  makeLiteralRule('table-header-group'),
-  makeLiteralRule('table-row'),
-  makeLiteralRule('table-row-group'),
-  makeLiteralRule('flex'),
-  makeLiteralRule('inline-flex'),
-  makeLiteralRule('grid'),
-  makeLiteralRule('inline-grid'),
-  makeLiteralRule('-webkit-box'),
-  makeLiteralRule('run-in'),
-  makeLiteralRule('ruby'),
-  makeLiteralRule('ruby-base'),
-  makeLiteralRule('ruby-text'),
-  makeLiteralRule('ruby-base-container'),
-  makeLiteralRule('ruby-text-container'),
-  makeLiteralRule('contents'),
-);
-const displayInside = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('block'),
-  makeLiteralRule('table'),
-  makeLiteralRule('flex'),
-  makeLiteralRule('grid'),
-  makeLiteralRule('ruby'),
-);
-const displayList = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('list-item'),
-);
-const displayOutside = makeUnionRule(
-  makeLiteralRule('block-level'),
-  makeLiteralRule('inline-level'),
-  makeLiteralRule('run-in'),
-  makeLiteralRule('contents'),
-  makeLiteralRule('none'),
-  makeLiteralRule('table-row-group'),
-  makeLiteralRule('table-header-group'),
-  makeLiteralRule('table-footer-group'),
-  makeLiteralRule('table-row'),
-  makeLiteralRule('table-cell'),
-  makeLiteralRule('table-column-group'),
-  makeLiteralRule('table-column'),
-  makeLiteralRule('table-caption'),
-  makeLiteralRule('ruby-base'),
-  makeLiteralRule('ruby-text'),
-  makeLiteralRule('ruby-base-container'),
-  makeLiteralRule('ruby-text-container'),
-);
-const fontFeatureSettings = makeUnionRule(makeLiteralRule('normal'), isString);
-const fontKerning = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('normal'),
-  makeLiteralRule('none'),
-);
-const fontLanguageOverride = makeUnionRule(makeLiteralRule('normal'), isString);
-const fontSize = makeUnionRule(absoluteSize, relativeSize, lengthPercentage);
-const fontSizeAdjust = makeUnionRule(
-  makeLiteralRule('none'),
-  isNumber,
-  isString,
-);
-const fontStretch = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('ultra-condensed'),
-  makeLiteralRule('extra-condensed'),
-  makeLiteralRule('condensed'),
-  makeLiteralRule('semi-condensed'),
-  makeLiteralRule('semi-expanded'),
-  makeLiteralRule('expanded'),
-  makeLiteralRule('extra-expanded'),
-  makeLiteralRule('ultra-expanded'),
-);
-const fontStyle = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('italic'),
-  makeLiteralRule('oblique'),
-);
-const fontSynthesis = makeUnionRule(makeLiteralRule('none'), isString);
-const fontVariant = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('none'),
-  isString,
-);
-const fontVariantAlternates = makeUnionRule(
-  makeLiteralRule('normal'),
-  isString,
-);
-const fontVariantCaps = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('small-caps'),
-  makeLiteralRule('all-small-caps'),
-  makeLiteralRule('petite-caps'),
-  makeLiteralRule('all-petite-caps'),
-  makeLiteralRule('unicase'),
-  makeLiteralRule('titling-caps'),
-);
-const fontVariantEastAsian = makeUnionRule(makeLiteralRule('normal'), isString);
-const fontVariantLigatures = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('none'),
-  isString,
-);
-const fontVariantNumeric = makeUnionRule(makeLiteralRule('normal'), isString);
-const fontVariantPosition = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('sub'),
-  makeLiteralRule('super'),
-);
-const fontWeight = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('bold'),
-  makeLiteralRule('bolder'),
-  makeLiteralRule('lighter'),
-  makeRangeRule(1, 1000, 'a number between 1 and 1000'),
-  isCSSVariable,
-);
-const gap = length;
-const grid = makeUnionRule(gridTemplate, isString);
-const gridArea = makeUnionRule(gridLine, isString);
-const gridAutoColumns = trackSize;
-const gridAutoFlow = makeUnionRule(isString, makeLiteralRule('dense'));
-const gridAutoRows = trackSize;
-const gridColumn = makeUnionRule(gridLine, isString);
-const gridColumnEnd = gridLine;
-const gridColumnStart = gridLine;
-const gridRow = makeUnionRule(gridLine, isString);
-const gridRowEnd = gridLine;
-const gridRowStart = gridLine;
-const hyphens = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('manual'),
-  makeLiteralRule('auto'),
-);
-const imageOrientation = makeUnionRule(
-  makeLiteralRule('from-image'),
-  isNumber,
-  isString,
-);
-const imageRendering = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('crisp-edges'),
-  makeLiteralRule('pixelated'),
-  makeLiteralRule('optimizeSpeed'),
-  makeLiteralRule('optimizeQuality'),
-  isString,
-);
-const imageResolution = makeUnionRule(isString, makeLiteralRule('snap'));
-const imeMode = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('normal'),
-  makeLiteralRule('active'),
-  makeLiteralRule('inactive'),
-  makeLiteralRule('disabled'),
-);
-const initialLetter = makeUnionRule(makeLiteralRule('normal'), isString);
-const initialLetterAlign = isString;
-const inlineSize = width;
-const interpolateSize = makeUnionRule('allow-keywords', 'numeric-only');
-const isolation = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('isolate'),
-);
-
-const justifyBase = [
-  'normal',
-  'stretch',
-  'center',
-  'start',
-  'end',
-  'flex-start',
-  'flex-end',
-  'left',
-  'right',
-  'baseline',
-  'first baseline',
-  'last baseline',
-  'safe center',
-  'unsafe center',
-];
-
-const justifyContent = makeUnionRule(
-  ...justifyBase,
-  'space-between',
-  'space-around',
-  'space-evenly',
-);
-
-const justifyItems = makeUnionRule(
-  ...justifyBase,
-  'self-start',
-  'self-end',
-  'legacy right',
-  'legacy left',
-  'legacy center',
-);
-
-// There's an optional overflowPosition (safe vs unsafe) prefix to
-// [selfPosition | 'left' | 'right']. It's not used on www, so, it's not added
-// here.
-const justifySelf = makeUnionRule(
-  'auto',
-  ...justifyBase,
-  'self-start',
-  'self-end',
-);
-
-const letterSpacing = makeUnionRule(
-  makeLiteralRule('normal'),
-  lengthPercentage,
-);
-const lineBreak = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('loose'),
-  makeLiteralRule('normal'),
-  makeLiteralRule('strict'),
-);
-const lineHeight = length;
-const listStyleImage = makeUnionRule(isString, makeLiteralRule('none'));
-const listStylePosition = makeUnionRule(
-  makeLiteralRule('inside'),
-  makeLiteralRule('outside'),
-);
-const listStyle = makeUnionRule(
-  listStyleType,
-  listStylePosition,
-  listStyleImage,
-);
-const margin = length;
-const marginLeft = makeUnionRule(
-  isNumber,
-  isNonNumericString,
-  makeLiteralRule('auto'),
-);
-const marginTop = makeUnionRule(
-  isNumber,
-  isNonNumericString,
-  makeLiteralRule('auto'),
-);
-const markerOffset = makeUnionRule(isNumber, makeLiteralRule('auto'));
-const mask = maskLayer;
-const maskClip = isString;
-const maskComposite = compositeOperator;
-const maskMode = maskingMode;
-const maskOrigin = geometryBox;
-const maskPosition = isString;
-const maskRepeat = repeatStyle;
-const maskSize = bgSize;
-const maskType = makeUnionRule(
-  makeLiteralRule('luminance'),
-  makeLiteralRule('alpha'),
-);
-const minMaxLength = makeUnionRule(
-  isNumber,
-  isNonNumericString,
-  makeLiteralRule('none'),
-  makeLiteralRule('max-content'),
-  makeLiteralRule('min-content'),
-  makeLiteralRule('fit-content'),
-  makeLiteralRule('fill-available'),
-);
-
-const mixBlendMode = blendMode;
-const motionPath = makeUnionRule(
-  isString,
-  geometryBox,
-  makeLiteralRule('none'),
-);
-const motionRotation = isStringOrNumber;
-const motion = makeUnionRule(motionPath, lengthPercentage, motionRotation);
-const objectFit = makeUnionRule(
-  makeLiteralRule('fill'),
-  makeLiteralRule('contain'),
-  makeLiteralRule('cover'),
-  makeLiteralRule('none'),
-  makeLiteralRule('scale-down'),
-);
-const objectPosition = isString;
-const offsetBlockEnd = isString;
-const offsetBlockStart = isString;
-const offsetInlineEnd = isString;
-const offsetInlineStart = isString;
-const opacity = isStringOrNumber;
-const order = isStringOrNumber;
-const orphans = isStringOrNumber;
-const outline = isString;
-const overflow = makeUnionRule(
-  makeLiteralRule('visible'),
-  makeLiteralRule('hidden'),
-  makeLiteralRule('clip'),
-  makeLiteralRule('scroll'),
-  makeLiteralRule('auto'),
-);
-const overflowAnchor = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('none'),
-);
-const overflowClipBox = makeUnionRule(
-  makeLiteralRule('padding-box'),
-  makeLiteralRule('content-box'),
-);
-const overflowWrap = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('break-word'),
-  makeLiteralRule('anywhere'),
-);
-const overflowDir = makeUnionRule(
-  makeLiteralRule('visible'),
-  makeLiteralRule('hidden'),
-  makeLiteralRule('scroll'),
-  makeLiteralRule('auto'),
-);
-const overscrollBehavior = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('contain'),
-  makeLiteralRule('auto'),
-);
-
-const pageBreak = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('always'),
-  makeLiteralRule('avoid'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-);
-const pageBreakInside = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('avoid'),
-);
-const perspective = makeUnionRule(makeLiteralRule('none'), isNumber, isString);
-const perspectiveOrigin = isString;
-const pointerEvents = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('none'),
-  makeLiteralRule('visiblePainted'),
-  makeLiteralRule('visibleFill'),
-  makeLiteralRule('visibleStroke'),
-  makeLiteralRule('visible'),
-  makeLiteralRule('painted'),
-  makeLiteralRule('fill'),
-  makeLiteralRule('stroke'),
-  makeLiteralRule('all'),
-);
-const position = makeUnionRule(
-  makeLiteralRule('static'),
-  makeLiteralRule('relative'),
-  makeLiteralRule('absolute'),
-  makeLiteralRule('sticky'),
-  makeLiteralRule('fixed'),
-);
-const quotes = makeUnionRule(isString, makeLiteralRule('none'));
-const resize = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('both'),
-  makeLiteralRule('horizontal'),
-  makeLiteralRule('vertical'),
-);
-const rowGap = length;
-const rubyAlign = makeUnionRule(
-  makeLiteralRule('start'),
-  makeLiteralRule('center'),
-  makeLiteralRule('space-between'),
-  makeLiteralRule('space-around'),
-);
-const rubyMerge = makeUnionRule(
-  makeLiteralRule('separate'),
-  makeLiteralRule('collapse'),
-  makeLiteralRule('auto'),
-);
-const rubyPosition = makeUnionRule(
-  makeLiteralRule('over'),
-  makeLiteralRule('under'),
-  makeLiteralRule('inter-character'),
-);
-const scrollBehavior = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('smooth'),
-);
-const scrollSnapPaddingBottom = isStringOrNumber;
-const scrollSnapPaddingTop = isStringOrNumber;
-const scrollSnapAlign = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('start'),
-  makeLiteralRule('end'),
-  makeLiteralRule('center'),
-);
-const scrollSnapType = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('x mandatory'),
-  makeLiteralRule('y mandatory'),
-);
-const shapeImageThreshold = isStringOrNumber;
-const shapeOutside = makeUnionRule(makeLiteralRule('none'), shapeBox, isString);
-const tabSize = isStringOrNumber;
-const tableLayout = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('fixed'),
-);
-const textAlign = makeUnionRule(
-  makeLiteralRule('start'),
-  makeLiteralRule('end'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('center'),
-  makeLiteralRule('justify'),
-  makeLiteralRule('match-parent'),
-);
-const textAlignLast = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('start'),
-  makeLiteralRule('end'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  makeLiteralRule('center'),
-  makeLiteralRule('justify'),
-);
-const textCombineUpright = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('all'),
-  isString,
-);
-const textDecorationColor = color;
-const textDecorationLine = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('underline'),
-  makeLiteralRule('overline'),
-  makeLiteralRule('line-through'),
-  makeLiteralRule('blink'),
-  isString,
-);
-// const textDecorationSkip = makeUnionRule(makeLiteralRule('none'), isString);
-const textDecorationStyle = makeUnionRule(
-  makeLiteralRule('solid'),
-  makeLiteralRule('double'),
-  makeLiteralRule('dotted'),
-  makeLiteralRule('dashed'),
-  makeLiteralRule('wavy'),
-);
-const textDecoration = makeUnionRule(
-  textDecorationLine,
-  textDecorationStyle,
-  textDecorationColor,
-);
-const textEmphasisColor = color;
-const textEmphasisPosition = isString;
-const textEmphasisStyle = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('filled'),
-  makeLiteralRule('open'),
-  makeLiteralRule('dot'),
-  makeLiteralRule('circle'),
-  makeLiteralRule('double-circle'),
-  makeLiteralRule('triangle'),
-  makeLiteralRule('filled sesame'),
-  makeLiteralRule('open sesame'),
-  isString,
-);
-const textEmphasis = makeUnionRule(textEmphasisStyle, textEmphasisColor);
-const textIndent = makeUnionRule(
-  lengthPercentage,
-  makeLiteralRule('hanging'),
-  makeLiteralRule('each-line'),
-);
-const textOrientation = makeUnionRule(
-  makeLiteralRule('mixed'),
-  makeLiteralRule('upright'),
-  makeLiteralRule('sideways'),
-);
-const textOverflow = makeUnionRule(
-  makeLiteralRule('clip'),
-  makeLiteralRule('ellipsis'),
-  isString,
-);
-const textRendering = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('optimizeSpeed'),
-  makeLiteralRule('optimizeLegibility'),
-  makeLiteralRule('geometricPrecision'),
-);
-const textShadow = makeUnionRule(makeLiteralRule('none'), isString);
-const textSizeAdjust = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('auto'),
-  isString,
-);
-const textTransform = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('capitalize'),
-  makeLiteralRule('uppercase'),
-  makeLiteralRule('lowercase'),
-  makeLiteralRule('full-width'),
-);
-const textUnderlinePosition = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('under'),
-  makeLiteralRule('left'),
-  makeLiteralRule('right'),
-  isString,
-);
-const textUnderlineOffset = makeUnionRule(
-  makeLiteralRule('auto'),
-  isLength,
-  isPercentage,
-);
-const touchAction = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('none'),
-  isString,
-  makeLiteralRule('manipulation'),
-);
-const transform = makeUnionRule(makeLiteralRule('none'), isString);
-const transformBox = makeUnionRule(
-  makeLiteralRule('border-box'),
-  makeLiteralRule('fill-box'),
-  makeLiteralRule('view-box'),
-  makeLiteralRule('content-box'),
-  makeLiteralRule('stroke-box'),
-);
-const transformOrigin = isStringOrNumber;
-const transformStyle = makeUnionRule(
-  makeLiteralRule('flat'),
-  makeLiteralRule('preserve-3d'),
-);
-const transitionProperty = isString;
-const transitionTimingFunction = animationTimingFunction;
-const unicodeBidi = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('embed'),
-  makeLiteralRule('isolate'),
-  makeLiteralRule('bidi-override'),
-  makeLiteralRule('isolate-override'),
-  makeLiteralRule('plaintext'),
-);
-const userSelect = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('text'),
-  makeLiteralRule('none'),
-  makeLiteralRule('contain'),
-  makeLiteralRule('all'),
-);
-const verticalAlign = makeUnionRule(
-  makeLiteralRule('baseline'),
-  makeLiteralRule('sub'),
-  makeLiteralRule('super'),
-  makeLiteralRule('text-top'),
-  makeLiteralRule('text-bottom'),
-  makeLiteralRule('middle'),
-  makeLiteralRule('top'),
-  makeLiteralRule('bottom'),
-  isString,
-  isNumber,
-);
-const visibility = makeUnionRule(
-  makeLiteralRule('visible'),
-  makeLiteralRule('hidden'),
-  makeLiteralRule('collapse'),
-);
-const whiteSpace = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('pre'),
-  makeLiteralRule('nowrap'),
-  makeLiteralRule('pre-wrap'),
-  makeLiteralRule('pre-line'),
-  makeLiteralRule('break-spaces'),
-);
-const widows = isStringOrNumber;
-const animatableFeature = makeUnionRule(
-  makeLiteralRule('scroll-position'),
-  makeLiteralRule('contents'),
-  isString,
-);
-const willChange = makeUnionRule(makeLiteralRule('auto'), animatableFeature);
-const nonStandardWordBreak = makeLiteralRule('break-word');
-const wordBreak = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('break-all'),
-  makeLiteralRule('keep-all'),
-  nonStandardWordBreak,
-);
-const wordSpacing = makeUnionRule(makeLiteralRule('normal'), lengthPercentage);
-const wordWrap = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('break-word'),
-);
-const svgWritingMode = makeUnionRule(
-  makeLiteralRule('lr-tb'),
-  makeLiteralRule('rl-tb'),
-  makeLiteralRule('tb-rl'),
-  makeLiteralRule('lr'),
-  makeLiteralRule('rl'),
-  makeLiteralRule('tb'),
-);
-const writingMode = makeUnionRule(
-  makeLiteralRule('horizontal-tb'),
-  makeLiteralRule('vertical-rl'),
-  makeLiteralRule('vertical-lr'),
-  makeLiteralRule('sideways-rl'),
-  makeLiteralRule('sideways-lr'),
-  svgWritingMode,
-);
-const zIndex = makeUnionRule(makeLiteralRule('auto'), isNumber);
-const alignmentBaseline = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('baseline'),
-  makeLiteralRule('before-edge'),
-  makeLiteralRule('text-before-edge'),
-  makeLiteralRule('middle'),
-  makeLiteralRule('central'),
-  makeLiteralRule('after-edge'),
-  makeLiteralRule('text-after-edge'),
-  makeLiteralRule('ideographic'),
-  makeLiteralRule('alphabetic'),
-  makeLiteralRule('hanging'),
-  makeLiteralRule('mathematical'),
-);
-const svgLength = isStringOrNumber;
-const baselineShift = makeUnionRule(
-  makeLiteralRule('baseline'),
-  makeLiteralRule('sub'),
-  makeLiteralRule('super'),
-  svgLength,
-);
-const behavior = isString;
-const clipRule = makeUnionRule(
-  makeLiteralRule('nonzero'),
-  makeLiteralRule('evenodd'),
-);
-const cueAfter = makeUnionRule(isStringOrNumber, makeLiteralRule('none'));
-const cueBefore = makeUnionRule(isStringOrNumber, makeLiteralRule('none'));
-const cue = makeUnionRule(cueBefore, cueAfter);
-const dominantBaseline = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('use-script'),
-  makeLiteralRule('no-change'),
-  makeLiteralRule('reset-size'),
-  makeLiteralRule('ideographic'),
-  makeLiteralRule('alphabetic'),
-  makeLiteralRule('hanging'),
-  makeLiteralRule('mathematical'),
-  makeLiteralRule('central'),
-  makeLiteralRule('middle'),
-  makeLiteralRule('text-after-edge'),
-  makeLiteralRule('text-before-edge'),
-);
-const paint = makeUnionRule(
-  makeLiteralRule('none'),
-  makeLiteralRule('currentColor'),
-  color,
-  isString,
-);
-const fill = paint;
-const fillOpacity = isStringOrNumber;
-const fillRule = makeUnionRule(
-  makeLiteralRule('nonzero'),
-  makeLiteralRule('evenodd'),
-);
-const glyphOrientationHorizontal = isStringOrNumber;
-const glyphOrientationVertical = isStringOrNumber;
-const kerning = makeUnionRule(makeLiteralRule('auto'), svgLength);
-const marker = makeUnionRule(makeLiteralRule('none'), isString);
-const markerEnd = makeUnionRule(makeLiteralRule('none'), isString);
-const markerMid = makeUnionRule(makeLiteralRule('none'), isString);
-const markerStart = makeUnionRule(makeLiteralRule('none'), isString);
-const pauseOrRest = makeUnionRule(
-  isNumber,
-  makeLiteralRule('none'),
-  makeLiteralRule('x-weak'),
-  makeLiteralRule('weak'),
-  makeLiteralRule('medium'),
-  makeLiteralRule('strong'),
-  makeLiteralRule('x-strong'),
-);
-const shapeRendering = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('optimizeSpeed'),
-  makeLiteralRule('crispEdges'),
-  makeLiteralRule('geometricPrecision'),
-);
-const src = isString;
-const speak = makeUnionRule(
-  makeLiteralRule('auto'),
-  makeLiteralRule('none'),
-  makeLiteralRule('normal'),
-);
-const speakAs = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('spell-out'),
-  makeLiteralRule('digits'),
-  isString,
-);
-const stroke = paint;
-const strokeDasharray = makeUnionRule(makeLiteralRule('none'), isString);
-const strokeDashoffset = svgLength;
-const strokeLinecap = makeUnionRule(
-  makeLiteralRule('butt'),
-  makeLiteralRule('round'),
-  makeLiteralRule('square'),
-);
-const strokeLinejoin = makeUnionRule(
-  makeLiteralRule('miter'),
-  makeLiteralRule('round'),
-  makeLiteralRule('bevel'),
-);
-const strokeMiterlimit = isStringOrNumber;
-const strokeOpacity = isStringOrNumber;
-const strokeWidth = svgLength;
-const textAnchor = makeUnionRule(
-  makeLiteralRule('start'),
-  makeLiteralRule('middle'),
-  makeLiteralRule('end'),
-);
-const unicodeRange = isString;
-const voiceBalance = makeUnionRule(
-  isNumber,
-  makeLiteralRule('left'),
-  makeLiteralRule('center'),
-  makeLiteralRule('right'),
-  makeLiteralRule('leftwards'),
-  makeLiteralRule('rightwards'),
-);
-const voiceDuration = makeUnionRule(makeLiteralRule('auto'), time);
-const voiceFamily = makeUnionRule(isString, makeLiteralRule('preserve'));
-const voicePitch = makeUnionRule(
-  isNumber,
-  makeLiteralRule('absolute'),
-  isString,
-);
-const voiceRange = makeUnionRule(
-  isNumber,
-  makeLiteralRule('absolute'),
-  isString,
-);
-const voiceRate = isString;
-const voiceStress = makeUnionRule(
-  makeLiteralRule('normal'),
-  makeLiteralRule('strong'),
-  makeLiteralRule('moderate'),
-  makeLiteralRule('none'),
-  makeLiteralRule('reduced'),
-);
-const voiceVolume = makeUnionRule(makeLiteralRule('silent'), isString);
-const maskImage = maskReference;
-
-const SupportedVendorSpecificCSSProperties = {
-  MozOsxFontSmoothing: makeLiteralRule('grayscale'),
-  WebkitFontSmoothing: makeLiteralRule('antialiased'),
-  WebkitAppearance: makeLiteralRule('textfield'),
-  WebkitTapHighlightColor: color,
-  WebkitOverflowScrolling: makeLiteralRule('touch'),
-
-  WebkitBoxOrient: makeUnionRule(
-    'horizontal',
-    'vertical',
-    'inline-axis',
-    'block-axis',
-  ),
-  WebkitLineClamp: isStringOrNumber,
-
-  WebkitMaskImage: maskImage,
-
-  WebkitTextFillColor: color,
-  textFillColor: color,
-  WebkitTextStrokeWidth: borderWidth,
-  WebkitTextStrokeColor: color,
-  WebkitBackgroundClip: makeUnionRule(
-    'border-box',
-    'padding-box',
-    'content-box',
-    'text',
-  ),
-};
-
-const convertToStandardProperties: $ReadOnly<{ [key: string]: ?string }> = {
-  marginStart: 'marginInlineStart',
-  marginEnd: 'marginInlineEnd',
-  marginHorizontal: 'marginInline',
-  marginVertical: 'marginBlock',
-
-  paddingStart: 'paddingInlineStart',
-  paddingEnd: 'paddingInlineEnd',
-  paddingHorizontal: 'paddingInline',
-  paddingVertical: 'paddingBlock',
-
-  borderVerticalWidth: 'borderBlockWidth',
-  borderVerticalStyle: 'borderBlockStyle',
-  borderVerticalColor: 'borderBlockColor',
-  borderHorizontalWidth: 'borderInlineWidth',
-  borderHorizontalStyle: 'borderInlineStyle',
-  borderHorizontalColor: 'borderInlineColor',
-  borderStartWidth: 'borderInlineStartWidth',
-  borderStartStyle: 'borderInlineStartStyle',
-  borderStartColor: 'borderInlineStartColor',
-  borderEndWidth: 'borderInlineEndWidth',
-  borderEndStyle: 'borderInlineEndStyle',
-  borderEndColor: 'borderInlineEndColor',
-
-  borderTopStartRadius: 'borderStartStartRadius',
-  borderTopEndRadius: 'borderStartEndRadius',
-  borderBottomStartRadius: 'borderEndStartRadius',
-  borderBottomEndRadius: 'borderEndEndRadius',
-
-  end: 'insetInlineEnd',
-  start: 'insetInlineStart',
-};
-
-const SVGProperties = {
-  colorInterpolation: makeUnionRule('auto', 'sRGB', 'linearRGB'),
-  // colorRendering: color,
-
-  fill: fill,
-  fillOpacity: fillOpacity,
-  fillRule: fillRule,
-
-  floodColor: color,
-  floodOpacity: opacity,
-
-  stopColor: color,
-  stopOpacity: opacity,
-
-  stroke: stroke,
-  strokeDasharray: strokeDasharray,
-  strokeDashoffset: strokeDashoffset,
-  strokeLinecap: strokeLinecap,
-  strokeLinejoin: strokeLinejoin,
-  strokeMiterlimit: strokeMiterlimit,
-  strokeOpacity: strokeOpacity,
-  strokeWidth: strokeWidth,
-
-  vectorEffect: makeUnionRule(
-    'none',
-    'non-scaling-stroke',
-    'non-scaling-size',
-    'non-rotation',
-    'fixed-position',
-  ),
-};
-
-/* eslint-disable object-shorthand */
-const CSSProperties = {
-  ...SupportedVendorSpecificCSSProperties,
-  ...SVGProperties,
-  accentColor: color,
-  alignTracks: isString,
-  alignContent: alignContent,
-  alignItems: alignItems,
-  alignSelf: alignSelf,
-  alignmentBaseline: alignmentBaseline,
-  all: all,
-  animation: showError(
-    '`animation` is not recommended. Please use `animationName`, `animationDuration`, etc. instead',
-  ),
-  animationComposition: makeUnionRule('replace', 'add', 'accumulate'),
-  animationDelay: time,
-  animationDirection: animationDirection,
-  animationDuration: time,
-  animationFillMode: animationFillMode,
-  animationIterationCount: animationIterationCount,
-  animationPlayState: animationPlayState,
-  animationTimingFunction: animationTimingFunction,
-  animationTimeline: isString,
-  appearance: appearance,
-  aspectRatio: isStringOrNumber,
-  backdropFilter: backdropFilter,
-  backfaceVisibility: backfaceVisibility,
-  background: isString,
-  backgroundAttachment: backgroundAttachment,
-  backgroundBlendMode: backgroundBlendMode,
-  backgroundClip: backgroundClip,
-  backgroundColor: backgroundColor,
-  backgroundImage: backgroundImage,
-  backgroundOrigin: backgroundOrigin,
-  backgroundPosition: backgroundPosition,
-  backgroundPositionX: backgroundPositionX,
-  backgroundPositionY: backgroundPositionY,
-  backgroundRepeat: backgroundRepeat,
-  backgroundSize: backgroundSize,
-  baselineShift: baselineShift,
-  behavior: behavior,
-
-  borderCollapse: borderCollapse,
-
-  borderImage: borderImage,
-  borderImageWidth: borderImageWidth,
-  borderImageOutset: borderImageOutset,
-  borderImageRepeat: borderImageRepeat,
-  borderImageSlice: borderImageSlice,
-  borderImageSource: borderImageSource,
-
-  borderWidth: borderWidth,
-  borderStyle: borderStyle,
-  borderColor: borderColor,
-  borderTopWidth: borderTopWidth,
-  borderTopStyle: borderTopStyle,
-  borderTopColor: color,
-  borderBottomWidth: borderBottomWidth,
-  borderBottomStyle: borderBottomStyle,
-  borderBottomColor: color,
-  borderLeftColor: borderLeftColor,
-  borderLeftStyle: borderLeftStyle,
-  borderLeftWidth: borderLeftWidth,
-  borderRightColor: borderLeftColor,
-  borderRightStyle: borderLeftStyle,
-  borderRightWidth: borderLeftWidth,
-  borderInlineEnd: showError(
-    [
-      '`borderInlineEnd` is not supported. Please use',
-      '  - `borderInlineEndWidth`,',
-      '  - `borderInlineEndStyle` and',
-      '  - `borderInlineEndColor` instead',
-    ].join('\n'),
-  ),
-  borderInlineColor: borderLeftColor,
-  borderInlineStyle: borderLeftStyle,
-  borderInlineWidth: borderLeftWidth,
-  borderInlineEndColor: borderLeftColor,
-  borderInlineEndStyle: borderLeftStyle,
-  borderInlineEndWidth: borderLeftWidth,
-  borderInlineStart: showError(
-    [
-      '`borderInlineStart` is not supported. Please use',
-      '  - `borderInlineStartWidth`,',
-      '  - `borderInlineStartStyle` and',
-      '  - `borderInlineStartColor` instead',
-    ].join('\n'),
-  ),
-  borderInlineStartColor: borderLeftColor,
-  borderInlineStartStyle: borderLeftStyle,
-  borderInlineStartWidth: borderLeftWidth,
-  borderBlockEnd: showError(
-    [
-      '`borderBlockEnd` is not supported. Please use',
-      '  - `borderBlockEndWidth`,',
-      '  - `borderBlockEndStyle` and',
-      '  - `borderBlockEndColor` instead',
-    ].join('\n'),
-  ),
-  borderBlockColor: borderLeftColor,
-  borderBlockStyle: borderLeftStyle,
-  borderBlockWidth: borderLeftWidth,
-  borderBlockEndColor: borderLeftColor,
-  borderBlockEndStyle: borderLeftStyle,
-  borderBlockEndWidth: borderLeftWidth,
-  borderBlockStart: showError(
-    [
-      '`borderBlockStart` is not supported. Please use',
-      '  - `borderBlockStartWidth`,',
-      '  - `borderBlockStartStyle` and',
-      '  - `borderBlockStartColor` instead',
-    ].join('\n'),
-  ),
-  borderBlockStartColor: borderLeftColor,
-  borderBlockStartStyle: borderLeftStyle,
-  borderBlockStartWidth: borderLeftWidth,
-
-  borderSpacing: borderSpacing,
-
-  borderRadius: lengthPercentage,
-  borderStartStartRadius: lengthPercentage,
-  borderStartEndRadius: lengthPercentage,
-  borderEndStartRadius: lengthPercentage,
-  borderEndEndRadius: lengthPercentage,
-
-  borderTopLeftRadius: lengthPercentage,
-  borderTopRightRadius: lengthPercentage,
-  borderBottomLeftRadius: lengthPercentage,
-  borderBottomRightRadius: lengthPercentage,
-
-  boxAlign: boxAlign,
-  boxDecorationBreak: boxDecorationBreak,
-  boxDirection: boxDirection,
-  boxFlex: boxFlex,
-  boxFlexGroup: boxFlexGroup,
-  boxLines: boxLines,
-  boxOrdinalGroup: boxOrdinalGroup,
-  boxOrient: boxOrient,
-  boxShadow: boxShadow,
-  boxSizing: boxSizing,
-  boxSuppress: boxSuppress,
-  breakAfter: breakBeforeOrAfter,
-  breakBefore: breakBeforeOrAfter,
-  breakInside: breakInside,
-  captionSide: captionSide,
-  caretColor: color,
-  clear: clear,
-  clip: clip,
-  clipPath: clipPath,
-  clipRule: clipRule,
-  color: color,
-  colorAdjust: makeUnionRule('economy', 'exact'),
-  colorScheme: makeUnionRule('light', 'dark', 'light dark'),
-  columnCount: columnCount,
-  columnFill: columnFill,
-  columnGap: columnGap,
-  columnRule: columnRule,
-  columnRuleColor: columnRuleColor,
-  columnRuleStyle: columnRuleStyle,
-  columnRuleWidth: columnRuleWidth,
-  columnSpan: columnSpan,
-  columnWidth: columnWidth,
-  columns: columns,
-  contain: contain,
-  containIntrinsicSize: makeUnionRule(isNumber, isString),
-  containIntrinsicBlockSize: makeUnionRule(isNumber, isString),
-  containIntrinsicInlineSize: makeUnionRule(isNumber, isString),
-  containIntrinsicHeight: makeUnionRule(isNumber, isString),
-  containIntrinsicWidth: makeUnionRule(isNumber, isString),
-  containerType: makeUnionRule('normal', 'size', 'inline-size'),
-  containerName: isString,
-  content: content,
-  contentVisibility: makeUnionRule('visible', 'hidden', 'auto'),
-  counterIncrement: counterIncrement,
-  counterReset: counterReset,
-  counterSet: isString,
-  cue: cue,
-  cueAfter: cueAfter,
-  cueBefore: cueBefore,
-  cursor: cursor,
-  direction: direction,
-  display: display,
-  displayInside: displayInside,
-  displayList: displayList,
-  displayOutside: displayOutside,
-  dominantBaseline: dominantBaseline,
-  emptyCells: emptyCells,
-
-  filter: filter,
-  flex: isString,
-  flexBasis: flexBasis,
-  flexDirection: flexDirection,
-  flexFlow: flexFlow,
-  flexGrow: flexGrow,
-  flexShrink: flexShrink,
-  flexWrap: flexWrap,
-  float: float,
-  font: showError(
-    '`font` is not recommended. Please use `fontSize`, `fontFamily`, `fontStyle` etc. instead',
-  ),
-  fontFamily: fontFamily,
-  fontFeatureSettings: fontFeatureSettings,
-  fontKerning: fontKerning,
-  fontLanguageOverride: fontLanguageOverride,
-  fontOpticalSizing: makeUnionRule('auto', 'none'),
-  fontPalette: isString,
-  fontSize: fontSize,
-  fontSizeAdjust: fontSizeAdjust,
-  fontSmooth: makeUnionRule('auto', 'never', 'always', lengthPercentage),
-  fontStretch: fontStretch,
-  fontStyle: fontStyle,
-  fontSynthesis: fontSynthesis,
-  fontVariant: fontVariant,
-  fontVariantAlternates: fontVariantAlternates,
-  fontVariantCaps: fontVariantCaps,
-  fontVariantEastAsian: fontVariantEastAsian,
-  fontVariantEmoji: makeUnionRule('auto', 'text', 'emoji', 'unicode'),
-  fontVariantLigatures: fontVariantLigatures,
-  fontVariantNumeric: fontVariantNumeric,
-  fontVariantPosition: fontVariantPosition,
-  fontVariationSettings: isString,
-  fontWeight: fontWeight,
-  forcedColorAdjust: makeUnionRule('auto', 'none'),
-
-  fieldSizing: makeUnionRule('content', 'fixed'),
-
-  gap: gap,
-  glyphOrientationHorizontal: glyphOrientationHorizontal,
-  glyphOrientationVertical: glyphOrientationVertical,
-
-  grid: grid,
-  gridArea: gridArea,
-  gridAutoColumns: gridAutoColumns,
-  gridAutoFlow: gridAutoFlow,
-  gridAutoRows: gridAutoRows,
-  gridColumn: gridColumn,
-  gridColumnEnd: gridColumnEnd,
-  gridColumnGap: lengthPercentage,
-  gridColumnStart: gridColumnStart,
-
-  gridRow: gridRow,
-  gridRowEnd: gridRowEnd,
-  gridGap: lengthPercentage,
-  gridRowGap: lengthPercentage,
-
-  gridRowStart: gridRowStart,
-  gridTemplate: gridTemplate,
-  gridTemplateAreas: gridTemplateAreas,
-  gridTemplateColumns: gridTemplate,
-  gridTemplateRows: gridTemplate,
-
-  hangingPunctuation: makeUnionRule(
-    'none',
-    'first',
-    'last',
-    'allow-end',
-    'force-end',
-    isString,
-  ),
-  hyphenateCharacter: isString,
-  hyphens: hyphens,
-  imageOrientation: imageOrientation,
-  imageRendering: imageRendering,
-  imageResolution: imageResolution,
-  imeMode: imeMode,
-  initialLetter: initialLetter,
-  initialLetterAlign: initialLetterAlign,
-
-  inset: length,
-  top: length,
-  right: length,
-  bottom: length,
-  left: length,
-  insetBlock: length,
-  insetBlockStart: length,
-  insetBlockEnd: length,
-  insetInline: length,
-  insetInlineStart: length,
-  insetInlineEnd: length,
-
-  height: width,
-  width: width,
-  blockSize: blockSize,
-  inlineSize: inlineSize,
-
-  interpolateSize: interpolateSize,
-
-  maxHeight: minMaxLength,
-  maxWidth: minMaxLength,
-  maxBlockSize: minMaxLength,
-  maxInlineSize: minMaxLength,
-  minBlockSize: minMaxLength,
-  minHeight: minMaxLength,
-  minInlineSize: minMaxLength,
-  minWidth: minMaxLength,
-
-  isolation: isolation,
-  justifyContent: justifyContent,
-  justifyItems: justifyItems,
-  justifySelf: justifySelf,
-  // Not supported in any browser yet.
-  // justifyTracks: makeUnionRule(
-  //   'start',
-  //   'end',
-  //   'center',
-  //   'normal',
-  //   'space-between',
-  //   'space-around',
-  //   'space-evenly',
-  //   'stretch',
-  //   'left',
-  //   'right',
-  // ),
-  kerning: kerning,
-  letterSpacing: letterSpacing,
-  lineBreak: lineBreak,
-  lineHeight: lineHeight,
-  listStyle: listStyle,
-  listStyleImage: listStyleImage,
-  listStylePosition: listStylePosition,
-  listStyleType: listStyleType,
-
-  margin: margin,
-  marginBlock: marginLeft,
-  marginBlockEnd: marginLeft,
-  marginBlockStart: marginLeft,
-  marginBottom: marginLeft,
-  marginInline: marginLeft,
-  marginInlineEnd: marginLeft,
-  marginInlineStart: marginLeft,
-  marginLeft: marginLeft,
-  marginRight: marginLeft,
-  marginTop: marginTop,
-
-  marker: marker,
-  markerEnd: markerEnd,
-  markerMid: markerMid,
-  markerOffset: markerOffset,
-  markerStart: markerStart,
-  mask: mask,
-  maskBorderMode: makeUnionRule('alpha', 'luminance'),
-  maskBorderOutset: isString,
-  maskBorderRepeat: makeUnionRule('stretch', 'repeat', 'round', 'space'),
-  maskBorderSlice: isString,
-  maskBorderSource: isString,
-  maskBorderWidth: isString,
-  maskClip: maskClip,
-  maskComposite: maskComposite,
-  maskImage: maskImage,
-  maskMode: maskMode,
-  maskOrigin: maskOrigin,
-  maskPosition: maskPosition,
-  maskRepeat: maskRepeat,
-  maskSize: maskSize,
-  maskType: maskType,
-
-  mixBlendMode: mixBlendMode,
-  motion: motion,
-  motionOffset: lengthPercentage,
-  motionPath: motionPath,
-  motionRotation: motionRotation,
-  objectFit: objectFit,
-  objectPosition: objectPosition,
-
-  offsetAnchor: isString,
-  offsetPath: isString,
-  offsetDistance: width,
-  offsetBlockEnd: offsetBlockEnd,
-  offsetBlockStart: offsetBlockStart,
-  offsetInlineEnd: offsetInlineEnd,
-  offsetInlineStart: offsetInlineStart,
-  offsetRotate: isString,
-  opacity: opacity,
-  order: order,
-  orphans: orphans,
-  outline: outline,
-  outlineColor: color,
-  outlineOffset: isLength,
-  outlineStyle: makeUnionRule(
-    'auto',
-    'none',
-    'dotted',
-    'dashed',
-    'solid',
-    'double',
-    'groove',
-    'ridge',
-    'inset',
-    'outset',
-  ),
-  outlineWidth: isLength,
-  blockOverflow: overflow, // TODO - Add support to Babel Plugin
-  inlineOverflow: overflow, // TODO - Add support to Babel Plugin
-  overflow: overflow,
-  overflowAnchor: overflowAnchor,
-  overflowClipBox: overflowClipBox,
-  overflowWrap: overflowWrap,
-  overflowX: overflowDir,
-  overflowY: overflowDir,
-  overscrollBehavior: overscrollBehavior,
-  // Currently Unsupported
-  // overscrollBehaviorInline: overscrollBehaviorX,
-  overscrollBehaviorX: overscrollBehavior,
-  overscrollBehaviorY: overscrollBehavior,
-  // Currently Unsupported
-  // overscrollBehaviorBlock: overscrollBehaviorY,
-  overflowClipMargin: isString,
-
-  paintOrder: makeUnionRule('normal', 'fill', 'stroke', 'markers', isString),
-
-  padding: length,
-  paddingBlock: length,
-  paddingBlockEnd: length,
-  paddingBlockStart: length,
-  paddingBottom: length,
-  paddingInline: length,
-  paddingInlineEnd: length,
-  paddingInlineStart: length,
-  paddingLeft: length,
-  paddingRight: length,
-  paddingTop: length,
-
-  pageBreakAfter: pageBreak,
-  pageBreakBefore: pageBreak,
-  pageBreakInside: pageBreakInside,
-  pause: pauseOrRest,
-  pauseAfter: pauseOrRest,
-  pauseBefore: pauseOrRest,
-  perspective: perspective,
-  perspectiveOrigin: perspectiveOrigin,
-  pointerEvents: pointerEvents,
-  position: position,
-
-  // Shorthand not yet supported
-  placeContent: isString,
-  placeItems: isString,
-  placeSelf: isString,
-  printColorAdjust: makeUnionRule('economy', 'exact'),
-
-  quotes: quotes,
-  resize: resize,
-  rest: pauseOrRest,
-  restAfter: pauseOrRest,
-  restBefore: pauseOrRest,
-
-  rotate: isString, // angle
-  scale: makeUnionRule(isString, isNumber),
-  translate: makeUnionRule(isString, isNumber),
-
-  rowGap: rowGap,
-  rubyAlign: rubyAlign,
-  rubyMerge: rubyMerge,
-  rubyPosition: rubyPosition,
-
-  scrollbarColor: color,
-  scrollbarGutter: makeUnionRule('auto', 'stable', 'stable both-edges'),
-  scrollbarWidth: makeUnionRule('auto', 'thin', 'none'),
-
-  scrollBehavior: scrollBehavior,
-  scrollSnapPaddingBottom: scrollSnapPaddingBottom,
-  scrollSnapPaddingTop: scrollSnapPaddingTop,
-  scrollSnapAlign: scrollSnapAlign,
-  scrollSnapType: scrollSnapType,
-  scrollSnapStop: makeUnionRule('normal', 'always'),
-
-  // scrollMargin: makeUnionRule(isNumber, isString),
-  scrollMarginBlockEnd: makeUnionRule(isNumber, isString),
-  scrollMarginBlockStart: makeUnionRule(isNumber, isString),
-  scrollMarginBottom: makeUnionRule(isNumber, isString),
-  scrollMarginInlineEnd: makeUnionRule(isNumber, isString),
-  scrollMarginInlineStart: makeUnionRule(isNumber, isString),
-  scrollMarginLeft: makeUnionRule(isNumber, isString),
-  scrollMarginRight: makeUnionRule(isNumber, isString),
-  scrollMarginTop: makeUnionRule(isNumber, isString),
-  scrollPaddingBlockEnd: makeUnionRule(isNumber, isString),
-  scrollPaddingBlockStart: makeUnionRule(isNumber, isString),
-  scrollPaddingBottom: makeUnionRule(isNumber, isString),
-  scrollPaddingInlineEnd: makeUnionRule(isNumber, isString),
-  scrollPaddingInlineStart: makeUnionRule(isNumber, isString),
-  scrollPaddingLeft: makeUnionRule(isNumber, isString),
-  scrollPaddingRight: makeUnionRule(isNumber, isString),
-  scrollPaddingTop: makeUnionRule(isNumber, isString),
-  scrollSnapMarginBottom: makeUnionRule(isNumber, isString),
-  scrollSnapMarginLeft: makeUnionRule(isNumber, isString),
-  scrollSnapMarginRight: makeUnionRule(isNumber, isString),
-  scrollSnapMarginTop: makeUnionRule(isNumber, isString),
-
-  shapeImageThreshold: shapeImageThreshold,
-  shapeMargin: lengthPercentage,
-  shapeOutside: shapeOutside,
-  shapeRendering: shapeRendering,
-  speak: speak,
-  speakAs: speakAs,
-  src: src,
-
-  tabSize: tabSize,
-  tableLayout: tableLayout,
-  textAlign: textAlign,
-  textAlignLast: textAlignLast,
-  textAnchor: textAnchor,
-  textCombineUpright: textCombineUpright,
-
-  textDecoration: textDecoration,
-  textDecorationColor: color,
-  textDecorationLine: isString, // TODO: Stricter support in the future
-  textDecorationSkip: makeUnionRule(
-    'none',
-    'objects',
-    'spaces',
-    'leading-spaces',
-    'trailing-spaces',
-    'edges',
-    'box-decoration',
-  ),
-  textDecorationSkipInk: makeUnionRule('none', 'auto', 'all'),
-  textDecorationStyle: makeUnionRule(
-    'solid',
-    'double',
-    'dotted',
-    'dashed',
-    'wavy',
-  ),
-  textDecorationThickness: isLength,
-
-  textEmphasis: textEmphasis,
-  textEmphasisColor: textEmphasisColor,
-  textEmphasisPosition: textEmphasisPosition,
-  textEmphasisStyle: textEmphasisStyle,
-  textIndent: textIndent,
-  textOrientation: textOrientation,
-  textOverflow: textOverflow,
-  textRendering: textRendering,
-  textShadow: textShadow,
-  textSizeAdjust: textSizeAdjust,
-  textTransform: textTransform,
-  textUnderlineOffset: textUnderlineOffset,
-  textUnderlinePosition: textUnderlinePosition,
-  textWrap: makeUnionRule('wrap', 'nowrap', 'balance', 'pretty'),
-
-  touchAction: touchAction,
-  transform: transform,
-  transformBox: transformBox,
-  transformOrigin: transformOrigin,
-  transformStyle: transformStyle,
-  transition: isString,
-  transitionDelay: time,
-  transitionDuration: time,
-  transitionProperty: transitionProperty,
-  transitionTimingFunction: transitionTimingFunction,
-  unicodeBidi: unicodeBidi,
-  unicodeRange: unicodeRange,
-  userSelect: userSelect,
-  viewTransitionName: makeUnionRule(all, isString),
-  verticalAlign: verticalAlign,
-  visibility: visibility,
-  voiceBalance: voiceBalance,
-  voiceDuration: voiceDuration,
-  voiceFamily: voiceFamily,
-  voicePitch: voicePitch,
-  voiceRange: voiceRange,
-  voiceRate: voiceRate,
-  voiceStress: voiceStress,
-  voiceVolume: voiceVolume,
-  whiteSpace: whiteSpace,
-  widows: widows,
-  willChange: willChange,
-  wordBreak: wordBreak,
-  wordSpacing: wordSpacing,
-  wordWrap: wordWrap,
-  writingMode: writingMode,
-  zIndex: zIndex,
-
-  // Purposely not supported because it is not supported in Firefox.
-  zoom: makeUnionRule('normal', 'reset', isNumber, isPercentage),
-};
-const CSSPropertyKeys = Object.keys(CSSProperties);
-for (const key of CSSPropertyKeys) {
-  CSSProperties[key] = makeUnionRule(CSSProperties[key], all);
-}
-
-const CSSPropertyReplacements: { [key: string]: RuleCheck | void } = {
-  border: border(),
-  borderTop: border('Top'),
-  borderBlockStart: border('Top'),
-  borderEnd: border('InlineEnd'),
-  borderInlineEnd: border('InlineEnd'),
-  borderRight: border('Right'),
-  borderBottom: border('Bottom'),
-  borderBlockEnd: border('Bottom'),
-  borderStart: border('InlineStart'),
-  borderInlineStart: border('InlineStart'),
-  borderLeft: border('Left'),
-};
-
-const pseudoElements = makeUnionRule(
-  makeLiteralRule('::before'),
-  makeLiteralRule('::after'),
-  makeLiteralRule('::first-letter'),
-  makeLiteralRule('::first-line'),
-  makeLiteralRule('::selection'),
-  makeLiteralRule('::backdrop'),
-  makeLiteralRule('::marker'),
-  makeLiteralRule('::placeholder'),
-  makeLiteralRule('::spelling-error'),
-  makeLiteralRule('::grammar-error'),
-  makeLiteralRule('::cue'),
-  makeLiteralRule('::slotted'),
-  makeLiteralRule('::part'),
-  makeLiteralRule('::thumb'),
-  // For styling input[type=number]
-  makeLiteralRule('::-webkit-inner-spin-button'),
-  makeLiteralRule('::-webkit-outer-spin-button'),
-  // For styling input[type=search]
-  makeLiteralRule('::-webkit-search-decoration'),
-  makeLiteralRule('::-webkit-search-cancel-button'),
-  makeLiteralRule('::-webkit-search-results-button'),
-  makeLiteralRule('::-webkit-search-results-decoration'),
-  // For Scrollbars in Webkit and Chromium
-  makeLiteralRule('::-webkit-scrollbar'),
-  makeLiteralRule('::-webkit-scrollbar-button'),
-  makeLiteralRule('::-webkit-scrollbar-thumb'),
-  makeLiteralRule('::-webkit-scrollbar-track'),
-  makeLiteralRule('::-webkit-scrollbar-track-piece'),
-  makeLiteralRule('::-webkit-scrollbar-corner'),
-  makeLiteralRule('::-webkit-resizer'),
-  // For input ranges in Chromium
-  makeLiteralRule('::-webkit-slider-thumb'),
-  makeLiteralRule('::-webkit-slider-runnable-track'),
-  // For input ranges in Firefox
-  makeLiteralRule('::-moz-range-thumb'),
-  makeLiteralRule('::-moz-range-track'),
-  makeLiteralRule('::-moz-range-progress'),
-);
-
-const pseudoClassesAndAtRules = makeUnionRule(
-  makeLiteralRule(':first-child'),
-  makeLiteralRule(':last-child'),
-  makeLiteralRule(':only-child'),
-  makeLiteralRule(':nth-child'),
-  makeLiteralRule(':nth-of-type'),
-  makeLiteralRule(':hover'),
-  makeLiteralRule(':focus'),
-  makeLiteralRule(':focus-visible'),
-  makeLiteralRule(':focus-within'),
-  makeLiteralRule(':active'),
-  makeLiteralRule(':visited'),
-  makeLiteralRule(':disabled'),
-  makeRegExRule(/^@media/, 'a media query'),
-  makeRegExRule(/^@container/, 'a media query'),
-  makeRegExRule(/^@supports/, 'a supports query'),
-  makeLiteralRule('@starting-style'),
-);
-
-const allModifiers = makeUnionRule(pseudoElements, pseudoClassesAndAtRules);
 
 const stylexValidStyles = {
   meta: {
